refactor : finebook모달 위치 변경
- 기존 detailFine에 위치
- 불필요한 props 드릴링 발생하여 모달 위치를 detailController로 이동
